First Example Without Using Custom Events
This version uses a push strategy, where the dialog pushes information to the UI by invoking its addTextBox method. The main issue here is that the dialog must know the method to call in the affected object (ui), and what parameters to pass. Defining the situation is the responsibility of the affected object, and it's up to the dialog to adhere to that. In OOP terms, the dialog "must know the semantics of the UI". Also, if there were multiple objects interested in the event, the dialog would need to invoke each one's appropriate function