in pull method,UI must know the semantics of the dialog
This demo uses a pull strategy, where the UI pulls information from the dialog when the event occurs. Here the situation is reversed -- now the UI must know the semantics of the dialog. Also, if there were multiple objects interested in the event, each one would have fairly complicated code to register for event notifications